Déplacement de getInputComponent
